Both flash and session cookies working
To unset session cookies I was deleting and trying to set again, but the
newly set cookies were still set to expire.
When reverting to setting the cookie to "" in order to unset, the
problem seems to have been that the path was set such that the cookie
was not being recognize/accepted. Setting path to '/' seems to have
fixed this.
This commit contains a lot of debugging crud, but works, so can be
tidied up later.